feat(iqb/cache): implement transform to IQB dict (#46)
This diff extends `cache.CacheEntry` and provides means to convert a
pair of dataframes (download and upload) into the dictionary expected by
the IQB calculator.

With this functionality in place, we can now compute IQB scores directly
from the v1 cached data and some sweating.

Subsequent work will wire this code into the code we currently use by
default to produce IQB scores with this library.