refactor(pipeline): separate cache and pipeline code (#51)
This diff separates code dealing with the cache layout on disk and code
dealing with the pipeline proper.

They were already conceptually separated, but part of the same file
(which means more difficulties moving code around and more difficulties
using AI, since it can lead to larger context windows, especially for
testing, and makes it more complex to maintain or adapt the code over
time).