refactor(PipelineCacheEntry)!: return `Path` not `Path | None` (#52)
This diff refactors the accessors for getting the `data.parquet` and
`stats.json` file paths of `PipelineCacheEntry`.

Before we used to return `None` to imply that the file does not exist,
but that was backwards. Consider:

1. the returned `Path` has an `exist` method the caller can invoke

2. the caller may want to *write into* the `Path` if not exists

So, this refactoring unlocks a second use case: knowing the paths in
which to write, which, in turn, reduces coupling.

BREAKING CHANGE: renamed `PipelineCacheEntry` `data_path` to
`data_parquet_file_path` and `stats_path` to `stats_json_file_path` and
updated their return value to be just `Path`.