fix(queries): ensure we include the AS name (#68)
I figured that, if we are showing the available ASNs in an area it's
definitely nicer to also show their name.

Test the queries by rerunning the pipeline and generating new data that
is now available at https://github.com/m-lab/iqb/releases/tag/v0.1.0 as
assets attached to the release.