doc(prototype): document automatic deployments (#72)
This diff improves upon the existing documentation to spell out how
automatic deployments work and to recap the URLs where we are publishing
the prototype.

This is a nice milestone for us, since now we can move faster and have
the `main` branch automatically published.

We also have the option of testing with the `sandbox` environment, which
feels useful for collaboration.

Lastly, I didn't mention this in the docs because we're not ready to use
this feature yet, but tagging would also publish to the production
environment.