feat(pipeline): introduce RemoteCache (#85)
This diff introduces a `RemoteCache` protocol that one can optionally
provide to the `IQBPipeline.get_cache_entry` method using the optional
`remote_cache` key-value argument.

If the key-value argument is `None`, the code behaves as before: check
locally and otherwise invoke BigQuery.

Otherwise, before invoking BigQuery, we attempt to fetch the cache entry
from the `remote_cache` (e.g., a GCS bucket).

Realistically, the first implementation of this protocol will fetch
cache entries from GitHub releases, this allowing us to replace the
functional dependency on `./data/ghcache.py sync`.

Also, note that this diff proceeds incrementally. It does not ask the
question of how the functionality of `./data/ghcache.py scan` (and the
related manual upload to GitHub) would be integrated with the library.
We want to automate the fetching part first. Once that is done, we will
also consider the problem of writing to the remote cache.