feat: introduce the scripting package (#109)
This package contains helper code to write scripts using IQB. For now,
we have just added logging helpers to have prettier logs. We are using
rich, which was introduced in the previous pull request.

The intention is to incrementally lift code we have already written for
throwaway scripts and factor common patterns.

By doing this, we will reduce the amount of untested code and gradually
build up a better UX for CLI usage.

The right moment to do this is ~now since I am using the CLI to fetch
more data, thus I can dogfod my own code.